# app/src/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "Calibration": {
        "description": "One `(host clock, encoder clock)` observation.\n\nThe first pair fixes the video's zero -- ffmpeg's PTS starts when it\ncaptured its first frame, which is some unknown time after we spawned it\n(process start, X11 connection, window lookup, first grab), and guessing\nthat offset is the mistake that makes every seek wrong by a constant nobody\ncan see. The second pair, at stop, fixes the *rate*, which is a check rather\nthan a parameter.",
        "properties": {
          "host_wall_ms": {
            "description": "Milliseconds since the recorder's epoch -- the same epoch\n[`TickSample::wall_ms`] counts from.",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "out_time_ms": {
            "description": "ffmpeg's own `out_time_us`, in milliseconds.",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "host_wall_ms",
          "out_time_ms"
        ],
        "type": "object"
      },
...
      "ClientRun": {
        "description": "One client's answer to \"which run do your frames belong to\".",
        "properties": {
          "client": {
            "description": "The `N` of the `client<N>` directory.",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "run": {
            "description": "That directory's own run id, `null` when it has no readable sidecar --\n*unknown*, never *no match*, exactly as in `FramesManifest::run`.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "client",
          "run"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Error body returned by every failing endpoint.\n\n`status` is the HTTP status [`IntoResponse`] answers with. It is not part of\nthe JSON body — it would be redundant with the response line, and clients\ngenerated from the OpenAPI schema must not see a field the server never\nserializes — so it is `#[serde(skip)]`ped and defaults to 400 on both\nconstruction and deserialization.",
        "properties": {
          "code": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "running_job_id": {
            "description": "The job holding the single execution slot, on the `409` from\n`POST /api/v1/scripts/execute`. Skipped when absent so every other\nerror body is byte-for-byte what it was before this field existed --\na client that switched on the presence of a key would otherwise start\nseeing `\"running_job_id\": null` on unrelated failures.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "message",
          "code"
        ],
        "type": "object"
      },
...
      "FrameEntry": {
        "description": "One frame file exactly as it exists on disk. One entry per file found\nunder a client's frames directory -- nothing is computed, nothing is\ninferred about files that are not there.",
        "properties": {
          "bytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "camera": {
            "description": "The camera id, parsed from the filename, or `null` alongside `tick`\nwhen the name does not parse.",
            "type": [
              "string",
              "null"
            ]
          },
          "client": {
            "description": "Which `client<N>` directory this frame was captured by.",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "name": {
            "description": "The filename exactly as it appears on disk.",
            "type": "string"
          },
          "tick": {
            "description": "`game.tick` at capture, parsed from the filename, or `null` when the\nname does not fit the pattern. Always present as a key -- `null` says\n\"we looked and there is nothing to report\", not \"we didn't ask\".",
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "client",
          "name",
          "bytes"
        ],
        "type": "object"
      },
      "FramesManifest": {
        "description": "`GET /api/v1/frames` response.\n\n`clients` and `frames` together distinguish the three states a caller must\nnot conflate:\n- no workspace, or no `client<N>` directory at all: `clients` is empty --\n  the run has never happened.\n- a `client<N>` directory exists but has produced no frames yet: `clients`\n  names it, `frames` is empty -- capture has not run.\n- frames are present: both are non-empty.",
        "properties": {
          "client_runs": {
            "description": "What each client's own sidecar says, in `clients` order.\n\n`run` above answers \"which run is this manifest about\"; this answers\n\"which run does each client's directory belong to\". They differ exactly\nwhen a client sat out the current run and kept an older one's frames,\nand that is the case a consumer needs to see rather than have decided\nfor it -- the stale client is named, so its frames can be marked\ninstead of silently mixed in with current ones.",
            "items": {
              "$ref": "#/components/schemas/ClientRun"
            },
            "type": "array"
          },
          "clients": {
            "description": "Every `client<N>` directory discovered under the workspace, regardless\nof whether it has produced a frame yet.",
            "items": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            },
            "type": "array"
          },
          "frames": {
            "description": "One entry per file actually on disk, sorted by `(client, tick,\ncamera)` with unparsed entries last, so iteration order is contractual\nrather than filesystem order.",
            "items": {
              "$ref": "#/components/schemas/FrameEntry"
            },
            "type": "array"
          },
          "run": {
            "description": "The opaque run identifier from `frames/run.json`, or `null` when the\ncapture was started without one, the file is absent, or it could not be\nread.\n\n**Opaque here too.** The server does not parse, validate or interpret\nit; it reads one string and forwards it, exactly as the mod echoes it\nwithout examining it. A consumer compares it for equality and nothing\nelse.\n\n`null` means *unknown*, never *no match* — a consumer that treats a\nmissing id as a mismatch would refuse a perfectly good join, and one\nthat treats it as a match would assert something nobody established.\n\n`#[schema(required)]` because the **key is always on the wire**: there\nis no `skip_serializing_if`, so serde emits `\"run\": null`. Without this\nutoipa would publish it as optional purely because the Rust type is an\n`Option`, and a generated client would then have to distinguish \"absent\nfrom the document\" from \"absent as a fact\" — a distinction this server\nnever makes and no consumer should have to.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "clients",
          "frames",
          "run",
          "client_runs"
        ],
        "type": "object"
      },
...
      "TickRange": {
        "description": "The observed span of a recording's clock.",
        "properties": {
          "from": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "to": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "from",
          "to"
        ],
        "type": "object"
      },
      "TickSample": {
        "description": "One line of `ticks.jsonl`, and one element of `GET /api/v1/video/ticks`.\n\nDeliberately one type for both. The file *is* the wire format; a second\nstruct for the HTTP side is how two halves of a system come to disagree\nabout what a sample is.",
        "properties": {
          "k": {
            "$ref": "#/components/schemas/TickKind",
            "description": "Always on the wire, even for the ordinary [`TickKind::Sample`].\n\nThe design's example lines omit it, and dropping it would save ~11 bytes\na line -- 60 KB across a 45-minute run, against a recording measured in\nhundreds of megabytes. It is kept because this struct is also the HTTP\nwire shape, and a key that is sometimes absent has to be declared\noptional in `app/src/api/types.ts`, where \"absent\" and \"null\" then become\na distinction the client has to make and this format never intends. A\nline written *without* `k` still parses, via `default`, so every document\nthe design describes is still readable.\n\n`#[schema(required)]` because the **key is always on the wire**: there is\nno `skip_serializing_if`, so serde always emits it. Without this utoipa\nwould publish it as optional purely because `#[serde(default)]` lets a\n*reader* omit it, and a generated client would then have to handle an\nabsence this server never produces."
          },
          "reason": {
            "description": "Why, for a gap. Absent otherwise.",
            "type": [
              "string",
              "null"
            ]
          },
          "t": {
            "description": "`game.tick` at the observation.\n\nAbsent on a [`TickKind::Gap`] line, and only there: a gap records that\nnothing was observed, which is not a tick with an unknown value but the\nabsence of an observation altogether.",
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "w": {
            "description": "Milliseconds since the recorder's own epoch -- the same epoch\n`VideoRecord::calibration` measures the video's PTS against.\n\nStamped at the **midpoint of send and receive**, never at receipt. The\ntick is the value inside the game at the moment the command ran, which\nis somewhere between the two; receipt time carries a systematic late\nbias of one full round trip, while the midpoint bounds the error at\nhalf a round trip and centres it on zero.",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "w",
          "k"
        ],
        "type": "object"
      },
      "VideoManifest": {
        "description": "Everything a caller needs to know about one video directory, derived from a\ndirectory listing every time it is asked for.\n\n**One type for the live directory and the archived one.** They hold the same\nfour files and answer the same questions; two shapes would be two chances to\ndisagree about what a recording is.",
        "properties": {
          "bytes": {
            "description": "The recording's size on disk, `null` when there is no file. A caller can\ntell \"the encoder was started and produced nothing\" from \"no encoder was\nstarted\" by pairing this with `video`.",
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "run": {
            "description": "The opaque run identifier from `video/run.json`, or `null` when the\ndirectory has no readable sidecar.\n\n`null` means *unknown*, never *no match* -- the same rule\n`FramesManifest::run` follows. A consumer compares it for equality and\ndoes nothing else with it.",
            "type": [
              "string",
              "null"
            ]
          },
          "samples": {
            "description": "How many lines `ticks.jsonl` holds, gaps included.",
            "minimum": 0,
            "type": "integer"
          },
          "skipped": {
            "description": "Lines of `ticks.jsonl` that did not parse -- in practice the torn last\nline of a killed recorder. Reported rather than swallowed.",
            "minimum": 0,
            "type": "integer"
          },
          "tick_range": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TickRange",
                "description": "The span of ticks the clock actually observed, `null` when it observed\nnone.\n\nPublished so the viewer can size its axis without fetching thousands of\nsamples: `runTimeline.ts` feeds this into the `drawn` set that decides\nwhere the axis starts, in the frames' place for a video-only run."
              }
            ]
          },
          "video": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VideoRecord",
                "description": "`video.json`, or `null` when the recorder never wrote one (which for a\nlive directory means video was never asked for on this run)."
              }
            ]
          }
        },
        "required": [
          "run",
          "video",
          "bytes",
          "samples",
          "skipped",
          "tick_range"
        ],
        "type": "object"
      },
      "VideoRecord": {
        "description": "`video.json`: what the encoder was asked for, what it did, and how it ended.",
        "properties": {
          "calibration": {
            "description": "The calibration pairs, in order. Zero on a failed start, one on a\nrecording that never reached a clean stop, two normally.",
            "items": {
              "$ref": "#/components/schemas/Calibration"
            },
            "type": "array"
          },
          "ffmpeg_exit": {
            "description": "ffmpeg's exit code, `null` when it never ran or was killed by a signal.",
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "file": {
            "description": "The recording's filename inside the video directory.",
            "type": "string"
          },
          "fps": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "height": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "rate_ok": {
            "description": "Whether the two calibration pairs agree on the rate within\n[`RATE_TOLERANCE`]. `null` when there are not two pairs to compare --\n*unknown*, never \"fine\".",
            "type": [
              "boolean",
              "null"
            ]
          },
          "reason": {
            "description": "Why, for a `failed`, `died` or `stopped_low_disk` recording.",
            "type": [
              "string",
              "null"
            ]
          },
          "requested_height": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "requested_width": {
            "description": "The geometry that was *asked for*. Kept beside the observed one rather\nthan instead of it: a tiling compositor ignores a size request unless\nthe window is floated, and a mismatch is a warning on the run, not a\nfailure -- the video is still usable and still joins on ticks.",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "run": {
            "description": "The run this recording belongs to.",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/VideoStatus"
          },
          "width": {
            "description": "The geometry `xwininfo` **observed**, which is what was actually\ncaptured. Zero on a recording that never started.",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "window_id": {
            "description": "The X11 window that was filmed, `0x`-prefixed. `null` when none was\nresolved.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "run",
          "file",
          "width",
          "height",
          "requested_width",
          "requested_height",
          "fps",
          "status",
          "reason",
          "ffmpeg_exit",
          "calibration",
          "rate_ok",
          "window_id"
        ],
        "type": "object"
      },
      "VideoStatus": {
        "description": "Where a recording got to.",
        "enum": [
          "recording",
          "stopped",
          "killed",
          "died",
          "failed",
          "stopped_low_disk"
        ],
        "type": "string"
      },
      "VideoTicksResponse": {
        "description": "`GET /api/v1/video/ticks` response.\n\nThe whole clock, in file order. The viewer interpolates between the two\nnearest samples and **refuses to interpolate across a gap** -- which is why\nthe gap lines are in this list rather than filtered out of it.",
        "properties": {
          "samples": {
            "items": {
              "$ref": "#/components/schemas/TickSample"
            },
            "type": "array"
          },
          "skipped": {
            "description": "Lines that did not parse -- in practice the torn last line of a recorder\nthat was killed. Reported rather than swallowed, matching\n`EventsResponse.skipped`.",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "samples",
          "skipped"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "email": "arturh@arturh.de",
      "name": "Artur Hallmann"
    },
    "description": "factorio-bot HTTP API",
    "license": {
      "name": ""
    },
    "title": "factorio-bot",
    "version": "0.2.4-dev"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/frames": {
      "get": {
        "description": "Derived from a directory listing on every call -- see the module docs for\nwhy this is not computed from a start tick and a stride.",
        "operationId": "list_frames",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FramesManifest"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Returns the manifest of frames captured so far.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/frames/{client}/{name}": {
      "get": {
        "description": "The route carries no client segment, so `name` is tried against each\ndiscovered `client<N>` directory in turn (lowest client number first),\nanswering the first match. Each directory is bounds-checked with\n[`resolve_script_path`] -- the same traversal guard\n`manage::scripts` uses -- rather than a second, independent\nimplementation of it: this endpoint is unauthenticated, so that guard is\nwhat stands between an HTTP caller and the filesystem.\n\n**Addressed by `(client, name)`, not by `name` alone.** Per-bot cameras mean\n`client1` and `client2` both capture at tick 300 and both files are correct\nand different; they are distinguished by the directory they are in, which\nis exactly what [`FrameEntry::client`] reports. An earlier version of this\nroute searched the client directories in order and returned the first\nmatch, which made the manifest honest and the bytes not: two entries with\nthe same `name` resolved to one image, and the higher-numbered client's\nframe was unreachable. A scrubber showing two bots would have shown the\nsame picture twice and nothing would have said so.\n\nStill no uniquifier: nothing here renames a colliding file so both survive.\nThat would let a genuine same-client double-write live on disk\nindistinguishably from two legitimate frames. The two cases are different —\ntwo clients at one tick is expected and addressable; one client twice at\none tick is a defect and must not be given a home.",
        "operationId": "get_frame",
        "parameters": [
          {
            "description": "the client number, as reported by FrameEntry.client",
            "in": "path",
            "name": "client",
            "required": true,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "a frame's filename, as reported by GET /api/v1/frames",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {}
            },
            "description": "the frame's JPEG bytes"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Serves one frame's JPEG bytes.",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/v1/runs/{id}/video/file": {
      "get": {
        "description": "Cached `immutable`, unlike the live route: this run is over, so these bytes\nnever change. The live `<workspace>/video/video.mp4` is overwritten by the\nnext run at the same URL and must not be.",
        "operationId": "get_run_video_file",
        "parameters": [
          {
            "description": "the run id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "video/mp4": {}
            },
            "description": "the whole recording"
          },
          "206": {
            "content": {
              "video/mp4": {}
            },
            "description": "a byte range, for seeking"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "One archived recording's bytes, with range support.",
        "tags": [
          "Runs"
        ]
      }
    },
...
    "/api/v1/video": {
      "get": {
        "description": "Answers a manifest describing *nothing* rather than a 404 when no video was\nrecorded: \"this run had no video\" is the ordinary state -- video is opt-in\n-- and a 404 would make a caller unable to tell it from a broken route.",
        "operationId": "get_video",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoManifest"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "The manifest of the current run's recording.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/video/file": {
      "get": {
        "description": "**`Cache-Control: no-cache`, deliberately unlike a frame.** A frame is\nimmutable once written -- a tick never recurs -- but\n`<workspace>/video/video.mp4` is *overwritten by the next run*, at the same\nURL. Serving it `immutable` would show a browser the previous run's\nrecording beside this run's timeline, and nothing would say so. The archived\nroute (`/api/v1/runs/{id}/video/file`) is the one whose bytes never change,\nand it is cached accordingly.",
        "operationId": "get_video_file",
        "responses": {
          "200": {
            "content": {
              "video/mp4": {}
            },
            "description": "the whole recording"
          },
          "206": {
            "content": {
              "video/mp4": {}
            },
            "description": "a byte range, for seeking"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Serves the recording's bytes, with range support.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/video/ticks": {
      "get": {
        "operationId": "get_video_ticks",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoTicksResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "The clock: every `(tick, wall_ms)` pair the recorder sampled.",
        "tags": [
          "Admin"
        ]
      }
    }
  },
...
}